chore: Extract some component from MC page
Add translations
