+++ src/locales/en-GB/messages.js
@@ -1,1 +1,1 @@
-/*eslint-disable*/module.exports={messages:JSON.parse("{\"zG7tgA\":\" Evil Lancelot thumbs up\",\"g/Qjku\":\" Total \",\"dyDb01\":[\"，\",[\"numEvilOpenEyes\"],\" evil players have eyes opened\"],\"g1kfb4\":\" Merlin \",\"TpJGH/\":\" Evil players have eyes opened \",\"IMhf1b\":\"Both Lancelot close their eyes\",\"J8eQpO\":\"Both Lancelot open their eyes\",\"ksg2O7\":\" and Morgana\",\"RSIg0B\":\" Evil players thumbs up, Merlin, opens your eyes \",\"TraFKg\":\"Everybody open your eyes\",\"cKXgmy\":\"All evil players close your eyes\",\"hANQJw\":\"All evil players open your eyes\",\"cLIbNJ\":\" thumbs down. Percival close your eyes\",\"SMB8c9\":\" Merlin close your eyes. Evil players thumbs down\",\"VmdlSD\":\"Total number of players\",\"PTfNJc\":\" thumbs up, Percival open your eyes\",\"cCIe39\":\"Except Oberon and Evil Lancelot,\",\"/aTJqV\":\"Except Evil Lancelot,\",\"jq6Vsj\":\"Except Oberon,\",\"bKCDxw\":\"Except Mordred\",\"new97P\":\" thumbs\"}")};
+/*eslint-disable*/module.exports={messages:JSON.parse("{\"zG7tgA\":\" Evil Lancelot thumbs up\",\"g/Qjku\":\" Total \",\"dyDb01\":[\"，\",[\"numEvilOpenEyes\"],\" evil players have eyes opened\"],\"g1kfb4\":\" Merlin \",\"ul0H4F\":\"Three, two, one.\",\"TpJGH/\":\" Evil players have eyes opened \",\"IMhf1b\":\"Both Lancelot close their eyes\",\"J8eQpO\":\"Both Lancelot open their eyes\",\"ksg2O7\":\" and Morgana\",\"9Gk1DL\":[\"Evil Players: \",[\"numberOfEvil\"]],\"RSIg0B\":\" Evil players thumbs up, Merlin, opens your eyes \",\"MrEN7u\":[\"Good Players: \",[\"numberOfGood\"]],\"TraFKg\":\"Everybody open your eyes\",\"cKXgmy\":\"All evil players close your eyes\",\"hANQJw\":\"All evil players open your eyes\",\"cLIbNJ\":\" thumbs down. Percival close your eyes\",\"SMB8c9\":\" Merlin close your eyes. Evil players thumbs down\",\"VmdlSD\":\"Total number of players\",\"B0dUAb\":[\"Total number of players: \",[\"totalNumberOfPlayer\"]],\"PTfNJc\":\" thumbs up, Percival open your eyes\",\"cCIe39\":\"Except Oberon and Evil Lancelot,\",\"/aTJqV\":\"Except Evil Lancelot,\",\"jq6Vsj\":\"Except Oberon,\",\"bKCDxw\":\"Except Mordred\",\"new97P\":\" thumbs\"}")};
