+++ src/locales/zh-hk/messages.js
@@ -1,1 +1,1 @@
-/*eslint-disable*/module.exports={messages:JSON.parse("{\"zG7tgA\":\"，壞蘭斯洛特豎起手指公\",\"g/Qjku\":\"，總共有\",\"dyDb01\":[\"，總共有\",[\"numEvilOpenEyes\"],\"個壞人開眼\"],\"g1kfb4\":\"【梅林】\",\"TpJGH/\":\"個壞人開眼\",\"IMhf1b\":\"兩個蘭斯洛特合埋眼\",\"J8eQpO\":\"兩個蘭斯洛特開眼確認身份\",\"ksg2O7\":\"及【莫甘娜】\",\"RSIg0B\":\"壞人豎起手指公，梅林擘大眼\",\"TraFKg\":\"所有人擘大眼\",\"cKXgmy\":\"所有壞人合埋眼\",\"hANQJw\":\"所有壞人擘大眼\",\"cLIbNJ\":\"收起手指公，【派西維爾】合埋眼。\",\"SMB8c9\":\"梅林合埋眼，壞人收起手指公\",\"VmdlSD\":\"總人數\",\"PTfNJc\":\"豎起手指公，【派西維爾】擘大眼。\",\"cCIe39\":\"除【奧柏倫同壞蘭斯洛特】以外，\",\"/aTJqV\":\"除左【壞蘭斯洛特】以外，\",\"jq6Vsj\":\"除左【奧柏倫】以外，\",\"bKCDxw\":\"除左【莫德雷德】以外，\",\"new97P\":\"隻手指公\"}")};
+/*eslint-disable*/module.exports={messages:JSON.parse("{\"zG7tgA\":\"，壞蘭斯洛特豎起手指公\",\"g/Qjku\":\"，總共有\",\"dyDb01\":[\"，總共有\",[\"numEvilOpenEyes\"],\"個壞人開眼\"],\"g1kfb4\":\"【梅林】\",\"ul0H4F\":\"三，二，一．\",\"TpJGH/\":\"個壞人開眼\",\"IMhf1b\":\"兩個蘭斯洛特合埋眼\",\"J8eQpO\":\"兩個蘭斯洛特開眼確認身份\",\"ksg2O7\":\"及【莫甘娜】\",\"9Gk1DL\":[\"壞人: \",[\"numberOfEvil\"],\" 個\"],\"RSIg0B\":\"壞人豎起手指公，梅林擘大眼\",\"MrEN7u\":[\"好人: \",[\"numberOfGood\"],\" 個\"],\"TraFKg\":\"所有人擘大眼\",\"cKXgmy\":\"所有壞人合埋眼\",\"hANQJw\":\"所有壞人擘大眼\",\"cLIbNJ\":\"收起手指公，【派西維爾】合埋眼。\",\"SMB8c9\":\"梅林合埋眼，壞人收起手指公\",\"VmdlSD\":\"總人數\",\"B0dUAb\":[\"總人數: \",[\"totalNumberOfPlayer\"]],\"PTfNJc\":\"豎起手指公，【派西維爾】擘大眼。\",\"cCIe39\":\"除【奧柏倫同壞蘭斯洛特】以外，\",\"/aTJqV\":\"除左【壞蘭斯洛特】以外，\",\"jq6Vsj\":\"除左【奧柏倫】以外，\",\"bKCDxw\":\"除左【莫德雷德】以外，\",\"new97P\":\"隻手指公\"}")};
